# services/competition-service/openapi/competition-service.v1.yaml
openapi: 3.0.3
info:
  title: competition-service API
  version: 0.4.0
  description: |
    Dịch vụ Thi đấu của Badminton Digital — chấm trình, điểm trình, bảng xếp hạng, xếp cặp.
    Bước 1: người chơi, chấm trình, BXH trình độ, thuật toán xếp cặp. Bước 2: trận đấu, giải đấu, BXH thành tích, thống kê.
    Bước 3: buổi giao lưu (điểm danh, xếp sân trống, màn hình lớn, đóng buổi).
    Plan 19: bấm điểm trực tiếp từng pha cầu, luồng SSE đẩy tỉ số lên màn hình TV.

    Xác thực: service token ES256 (gateway của app chính ký, sống ≤ 5 phút) — xem docs/01-kien-truc.md mục 6.
    Envelope: { success, data, message, errors } giống app chính, thêm `code` khi lỗi.
    Mọi POST có tác dụng phụ bắt buộc header Idempotency-Key.
servers:
  - url: /
security:
  - serviceToken: []
...
paths:
...
  /v1/events:
    post:
      tags: [ops]
      summary: Nhận sự kiện từ hệ thống khác (bd.customer.updated / merged / deleted)
      description: Xác thực bằng chữ ký HMAC (X-Event-Source, X-Timestamp, X-Signature), không dùng service token.
      security: [{ eventSignature: [] }]
      parameters:
        - { name: X-Event-Source, in: header, required: true, schema: { type: string } }
        - { name: X-Timestamp, in: header, required: true, schema: { type: string, pattern: '^[0-9]+$' } }
        - { name: X-Signature, in: header, required: true, schema: { type: string, pattern: '^sha256=[0-9a-f]{64}$' } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, type, data]
              properties:
                specversion: { type: string }
                id: { type: string, maxLength: 64 }
                type: { type: string, maxLength: 100 }
                source: { type: string }
                time: { type: string, format: date-time }
                data: { type: object }
      responses:
        '200':
          description: Đã nhận (kể cả trùng lặp)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [eventId, status]
                        properties:
                          eventId: { type: string }
                          status: { type: string, enum: [processed, ignored, duplicate] }
                          result: { type: string }
        default: { $ref: '#/components/responses/Error' }
  /v1/ops/outbox:
    get:
      tags: [ops]
      summary: Xem sự kiện chưa giao được (mặc định trạng thái dead)
      description: 'Scope: ops:admin'
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, delivered, dead, no_target] } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Paged' }
        default: { $ref: '#/components/responses/Error' }
  /v1/ops/outbox/{id}/replay:
    post:
      tags: [ops]
      summary: Gửi lại một sự kiện dead
      description: 'Scope: ops:admin'
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Envelope' }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/players/{id}/matches:
    parameters: [{ $ref: '#/components/parameters/PlayerId' }]
    get:
      tags: [match]
      summary: Lịch sử trận (đã chốt) hoặc trận sắp tới
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [history, upcoming] } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { $ref: '#/components/responses/MatchPage' }
        default: { $ref: '#/components/responses/Error' }
  /v1/players/{id}/partners:
    parameters: [{ $ref: '#/components/parameters/PlayerId' }]
    get:
      tags: [match]
      summary: 5 đồng đội hay đánh cặp nhất + tỉ lệ thắng khi đánh cùng
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [playerId, items]
                        properties:
                          playerId: { type: string, format: uuid }
                          items:
                            type: array
                            items:
                              type: object
                              required: [playerId, name, matches, wins, winRate]
                              properties:
                                playerId: { type: string, format: uuid }
                                name: { type: string, nullable: true }
                                matches: { type: integer }
                                wins: { type: integer }
                                winRate: { type: number }
        default: { $ref: '#/components/responses/Error' }
  /v1/players/{id}/head-to-head/{otherId}:
    parameters:
      - { $ref: '#/components/parameters/PlayerId' }
      - { name: otherId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [match]
      summary: Đối đầu hai người
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [playerId, otherId, matches, wins, losses, items]
                        properties:
                          playerId: { type: string, format: uuid }
                          otherId: { type: string, format: uuid }
                          matches: { type: integer }
                          wins: { type: integer }
                          losses: { type: integer }
                          items: { type: array, items: { $ref: '#/components/schemas/MatchView' } }
        default: { $ref: '#/components/responses/Error' }
  /v1/me/matches:
    get:
      tags: [match]
      summary: Trận của tôi (sắp tới / đã chốt)
      description: 'Scope: rating:self'
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [history, upcoming] } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200': { $ref: '#/components/responses/MatchPage' }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/matches/{id}:
    parameters: [{ $ref: '#/components/parameters/MatchId' }]
    get:
      tags: [match]
      summary: Chi tiết trận
      description: 'Scope đọc của giải / buổi, hoặc match:score + claim player của người chơi trong trận (màn hình bấm điểm).'
      responses:
        '200': { $ref: '#/components/responses/Match' }
        default: { $ref: '#/components/responses/Error' }
  /v1/matches/{id}/call:
    parameters: [{ $ref: '#/components/parameters/MatchId' }]
    post:
      tags: [match]
      summary: Gọi trận ra sân → in_play
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                courtRef: { type: string, maxLength: 64 }
      responses:
        '200': { $ref: '#/components/responses/Match' }
        default: { $ref: '#/components/responses/Error' }
  /v1/matches/{id}/result:
    parameters: [{ $ref: '#/components/parameters/MatchId' }]
    put:
      tags: [match]
      summary: Ghi / sửa kết quả trận (kiểm luật BWF; người thắng tự vào trận sau)
      description: 'Scope: tournament:operate (trận giải) hoặc session:operate (trận giao lưu). If-Match bắt buộc.'
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                games:
                  type: array
                  maxItems: 5
                  items: { type: array, minItems: 2, maxItems: 2, items: { type: integer, minimum: 0, maximum: 99 } }
                outcome: { type: string, enum: [normal, walkover, retired] }
                winnerSide: { type: string, enum: [A, B] }
      responses:
        '200': { $ref: '#/components/responses/Match' }
        default: { $ref: '#/components/responses/Error' }
  /v1/matches/{id}/cancel:
    parameters: [{ $ref: '#/components/parameters/MatchId' }]
    post:
      tags: [match]
      summary: Huỷ trận chưa có kết quả (không áp dụng cho trận trong sơ đồ; trận giao lưu bị huỷ không tính là đã đánh)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200': { $ref: '#/components/responses/Match' }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/matches/{id}/end:
    parameters: [{ $ref: '#/components/parameters/MatchId' }]
    post:
      tags: [match]
      summary: Kết thúc trận không nhập tỉ số (chỉ trận giao lưu) → ended, sân được nhả
      description: 'Scope: session:operate. Trận giải → 409 RESULT_REQUIRED. Nhập tỉ số sau đó vẫn được (ended → completed).'
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200': { $ref: '#/components/responses/Match' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT (ES256)
    eventSignature:
      type: apiKey
      in: header
      name: X-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Cùng key + cùng body → trả lại đúng response cũ; cùng key + khác body → 422.
      schema: { type: string, pattern: '^[A-Za-z0-9_-]{8,128}$' }
    PlayerId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    MatchId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: 'Version hiện tại (ETag). Ghi kết quả trận bắt buộc có — thiếu → 412, lệch → 409.'
      schema: { type: string, pattern: '^"?[0-9]+"?$' }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
  responses:
    Error:
      description: Lỗi (envelope có `code`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Match:
      description: Trận
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/MatchView' }
...
    MatchPage:
      description: Danh sách trận
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/Paged'
                      - type: object
                        properties:
                          items: { type: array, items: { $ref: '#/components/schemas/MatchView' } }
...
  schemas:
    Envelope:
      type: object
      required: [success, data, message, errors]
      properties:
        success: { type: boolean }
        data: { description: 'Dữ liệu trả về (null khi không có)' }
        message: { type: string, nullable: true }
        errors: { type: array, nullable: true, items: { $ref: '#/components/schemas/FieldError' } }
        code: { type: string }
    ErrorEnvelope:
      type: object
      required: [success, data, message, code, errors]
      properties:
        success: { type: boolean, enum: [false] }
        data: { description: 'Luôn null khi lỗi' }
        message: { type: string }
        code: { type: string }
        errors: { type: array, nullable: true, items: { $ref: '#/components/schemas/FieldError' } }
    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    Paged:
      type: object
      required: [items, total, page, totalPages]
      properties:
        items: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        totalPages: { type: integer }
...
    Discipline: { type: string, enum: [singles, doubles] }
...
    MatchTeam:
      type: object
      nullable: true
      required: [teamId, players]
      properties:
        teamId: { type: string, nullable: true, description: 'Đội của giải; null với trận giao lưu' }
        players:
          type: array
          items:
            type: object
            required: [id, name]
            properties:
              id: { type: string, format: uuid }
              name: { type: string, nullable: true }
    MatchLive:
      type: object
      nullable: true
      additionalProperties: false
      description: 'Bấm điểm trực tiếp (docs/06 mục 1.5) — tính lại từ chuỗi pha cầu. Trong MatchView: null = chưa ai bấm điểm.'
      required: [matchId, revision, rallies, firstServer, games, current, gameNo, gamesWon, server, serveFrom, decided, winnerSide, updatedAt]
      properties:
        matchId: { type: string, format: uuid }
        revision: { type: integer, minimum: 0, description: 'Tăng ở mọi lần đổi (kể cả hoàn tác) — gửi kèm khi bấm tiếp' }
        rallies: { type: string, pattern: '^[AB]*$', maxLength: 1000, description: 'Đội thắng từng pha cầu, theo thứ tự' }
        firstServer: { type: string, enum: [A, B] }
        games:
          type: array
          description: Các game đã xong
          items: { type: array, minItems: 2, maxItems: 2, items: { type: integer, minimum: 0 } }
        current:
          type: array
          nullable: true
          description: 'Game đang đánh [A, B]; null khi đã xong trận'
          minItems: 2
          maxItems: 2
          items: { type: integer, minimum: 0 }
        gameNo: { type: integer, minimum: 1 }
        gamesWon: { type: array, minItems: 2, maxItems: 2, items: { type: integer, minimum: 0 } }
        server: { type: string, nullable: true, enum: [A, B, null], description: 'Đội đang giao (đội thắng pha trước)' }
        serveFrom: { type: string, nullable: true, enum: [right, left, null], description: 'Điểm của đội giao chẵn → ô phải, lẻ → ô trái' }
        decided: { type: boolean, description: 'Đã đủ điểm thắng trận — chỉ còn xác nhận hoặc hoàn tác' }
        winnerSide: { type: string, nullable: true, enum: [A, B, null] }
        updatedAt: { type: string, format: date-time, nullable: true }
    MatchView: &matchView
      type: object
      additionalProperties: false
      required: [id, contextType, contextId, discipline, stage, teamA, teamB, scoring, games, outcome, winnerSide, status, live, version]
      properties:
        id: { type: string, format: uuid }
        contextType: { type: string, enum: [tournament, session] }
        contextId: { type: string, format: uuid }
        discipline: { $ref: '#/components/schemas/Discipline' }
        stage: { type: string, enum: [group, knockout, extra, session] }
        label: { type: string, nullable: true }
        groupNo: { type: integer, nullable: true }
        roundNo: { type: integer, nullable: true }
        slotNo: { type: integer, nullable: true }
        bracketPos: { type: integer, nullable: true }
        nextMatchId: { type: string, nullable: true }
        teamA: { $ref: '#/components/schemas/MatchTeam' }
        teamB: { $ref: '#/components/schemas/MatchTeam' }
        scoring: { $ref: '#/components/schemas/ScoringObject' }
        games: { type: array, items: { type: array, items: { type: integer } } }
        outcome: { type: string, nullable: true, enum: [normal, walkover, retired, null] }
        winnerSide: { type: string, nullable: true, enum: [A, B, null] }
        status: { type: string, enum: [scheduled, in_play, completed, cancelled, ended], description: 'ended = xong không nhập tỉ số (giao lưu)' }
        courtRef: { type: string, nullable: true }
        calledAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        live: { $ref: '#/components/schemas/MatchLive' }
        version: { type: integer }
    ScoringObject:
      type: object
      required: [bestOf, points]
      properties:
        bestOf: { type: integer, enum: [1, 3, 5] }
        points: { type: integer, minimum: 5, maximum: 50 }
        cap: { type: integer, nullable: true }
...